Fix animate_hero_car road_pos clamp: type, init, and comparison bugs
ahc_road_pos_b was declared u8, truncating the upper bound 472 ($01D8)
to 216 ($D8). This made the high-byte comparison always pass for any
road_pos >= 256, clamping it to 216 every frame.

The low-byte branch comparisons used (s8)A < 0 (sign check) instead of
the Z80's JR C (carry/borrow = unsigned L < E). This incorrectly
triggered for differences in [128,255], e.g. 216-72=144 appeared
negative.

The upper-bound high-byte equality test used A == 0 instead of
A == (DEother_road_pos >> 8), so road_pos with high byte matching the
bound's high byte always fell through to the clamp.

Both SM fields (default LD DE operands at $B395/$B3A3) were missing
initialisations in chq_initialise; calloc-zero is not the correct
default (72 and 472 are).

Net effect: road_pos 308 was clamped to 216, then 72; curvature
adjustment from move_hero_car pushed it back to 216 each frame,
producing the 72<->216 oscillation (road flicker / car bounce).

+++ C/ChaseHQ/ChaseHQ-State.h
@@ -529,7 +529,7 @@ struct chqstate {
   // $B395 (SM) in animate_hero_car
   u16       ahc_road_pos_a; // a road position
   // $B3A3 (SM) in animate_hero_car
-  u8        ahc_road_pos_b; // another road position
+  u16       ahc_road_pos_b; // another road position
   // $B3DB (SM) in animate_hero_car
   u8        ahc_crash_spin; // controls flipping
   // $B476 (SM) in animate_hero_car
